feat: change test, build config
