+++ .jest.js
@@ -1,44 +1,35 @@
 const libDir = process.env.LIB_DIR;
 
-const transformPackages = [
-  'react-native',
-  'react-native-menu',
-  'react-native-scrollable-tab-view',
-  'react-native-collapsible',
-  'react-native-swipeout',
-  'react-native-camera-roll-picker',
-]
+const transformIgnorePatterns = [
+  '/dist/',
+  'node_modules\/[^/]+?\/(?!es\/)', // Ignore modules without es dir
+];
 
 module.exports = {
-  preset: 'react-native',
-  setupFiles: [
-    './tests/setup.js'
-  ],
   moduleFileExtensions: [
     'ts',
     'tsx',
     'js',
     'jsx',
     'json',
+    'md',
   ],
   testPathIgnorePatterns: [
     '/node_modules/',
     '_site',
-    'site',
+    'site'
   ],
   transform: {
     '\\.tsx?$': './node_modules/antd-tools/lib/jest/codePreprocessor',
     '\\.js$': './node_modules/antd-tools/lib/jest/codePreprocessor',
-    '\\.png': '<rootDir>/tests/imageStub.js',
+    '\\.md$': './node_modules/antd-tools/lib/jest/demoPreprocessor',
+    '\\.svg$': '<rootDir>/tests/imageStub.js',
   },
   testRegex: libDir === 'dist' ? 'demo\\.test\\.js$' : '.*\\.test\\.js$',
   collectCoverageFrom: [
     'components/**/*.{ts,tsx}',
-    '!components/**/demo/*.{ts,tsx}',
-    '!components/**/*.web.{ts,tsx}',
-    '!components/*/style/index.tsx',
+    '!components/*/style/index.native.tsx',
   ],
-  transformIgnorePatterns: [
-    `node_modules/(?!(${transformPackages.join('|')})/)`,
-  ],
+  transformIgnorePatterns,
+  testEnvironment: 'jsdom',
 };
