+++ components/index.tsx
@@ -1,0 +1,67 @@
+// this file is not used if use https://github.com/ant-design/babel-plugin-import
+const ENV = process.env.NODE_ENV;
+if (ENV !== 'production' &&
+    ENV !== 'test' &&
+    typeof console !== 'undefined' &&
+    console.warn &&
+    typeof window !== 'undefined') {
+  console.warn(
+    'You are using a whole package of antd-mobile, ' +
+    'please use https://www.npmjs.com/package/babel-plugin-import to reduce app bundle size.',
+  );
+}
+
+export { default as Accordion } from './accordion/index.web';
+export { default as ActionSheet } from './action-sheet/index.web';
+export { default as ActivityIndicator } from './activity-indicator/index.web';
+export { default as Badge } from './badge/index.web';
+export { default as Button } from './button/index.web';
+export { default as Card } from './card/index.web';
+export { default as Carousel } from './carousel/index.web';
+export { default as Checkbox } from './checkbox/index.web';
+export { default as DatePicker } from './date-picker/index.web';
+export { default as Drawer } from './drawer/index.web';
+export { default as Flex } from './flex/index.web';
+export { default as Grid } from './grid/index.web';
+export { default as Icon } from './icon/index.web';
+export { default as ImagePicker } from './image-picker/index.web';
+export { default as InputItem } from './input-item/index.web';
+
+export { default as List } from './list/index.web';
+export { default as ListView } from './list-view/index.web';
+export { default as Menu } from './menu/index.web';
+export { default as Modal } from './modal/index.web';
+export { default as NavBar } from './nav-bar/index.web';
+export { default as NoticeBar } from './notice-bar/index.web';
+
+export { default as Pagination } from './pagination/index.web';
+export { default as Picker } from './picker/index.web';
+export { default as PickerView } from './picker-view/index.web';
+export { default as Popover } from './popover/index.web';
+export { default as Popup } from './popup/index.web';
+export { default as Progress } from './progress/index.web';
+
+export { default as Radio } from './radio/index.web';
+export { default as RefreshControl } from './refresh-control/index.web';
+export { default as Result } from './result/index.web';
+export { default as SearchBar } from './search-bar/index.web';
+export { default as SegmentedControl } from './segmented-control/index.web';
+export { default as Slider } from './slider/index.web';
+export { default as Range } from './range/index.web';
+export { default as createTooltip } from './create-tooltip/index.web';
+export { default as Stepper } from './stepper/index.web';
+export { default as Steps } from './steps/index.web';
+export { default as SwipeAction } from './swipe-action/index.web';
+export { default as Switch } from './switch/index.web';
+export { default as TabBar } from './tab-bar/index.web';
+export { default as Table } from './table/index.web';
+export { default as Tabs } from './tabs/index.web';
+export { default as Tag } from './tag/index.web';
+export { default as Text } from './text/index.web';
+export { default as TextareaItem } from './textarea-item/index.web';
+export { default as Toast } from './toast/index.web';
+export { default as View } from './view/index.web';
+export { default as WhiteSpace } from './white-space/index.web';
+export { default as WingBlank } from './wing-blank/index.web';
+
+export { default as LocaleProvider } from './locale-provider/index.web';
