+++ index.native.js
@@ -1,0 +1,44 @@
+// export components to react-native demo
+
+export { default as ActionSheet } from './components/action-sheet/';
+export { default as ActivityIndicator } from './components/activity-indicator/';
+export { default as Badge } from './components/badge/';
+export { default as Button } from './components/button/';
+export { default as Card } from './components/card/';
+export { default as Carousel } from './components/carousel/';
+export { default as Checkbox } from './components/checkbox/';
+export { default as DatePicker } from './components/date-picker/';
+export { default as Drawer } from './components/drawer/';
+export { default as Flex } from './components/flex/';
+export { default as Grid } from './components/grid';
+export { default as ImagePicker } from './components/image-picker/';
+export { default as InputItem } from './components/input-item/';
+export { default as List } from './components/list/';
+export { default as ListView } from './components/list-view/';
+export { default as Modal } from './components/modal/';
+
+export { default as NoticeBar } from './components/notice-bar/';
+export { default as Pagination } from './components/pagination/';
+export { default as Picker } from './components/picker/';
+export { default as Popover } from './components/popover/';
+export { default as Popup } from './components/popup/';
+export { default as Progress } from './components/progress/';
+export { default as Radio } from './components/radio/';
+export { default as RefreshControl } from './components/refresh-control/';
+export { default as Result } from './components/result/';
+export { default as SearchBar } from './components/search-bar/';
+export { default as SegmentedControl } from './components/segmented-control/';
+export { default as Slider } from './components/slider/';
+export { default as Stepper } from './components/stepper/';
+export { default as Steps } from './components/steps/';
+export { default as SwipeAction } from './components/swipe-action/';
+export { default as Switch } from './components/switch/';
+export { default as TabBar } from './components/tab-bar/';
+export { default as Tabs } from './components/tabs/';
+export { default as Tag } from './components/tag/';
+export { default as TextareaItem } from './components/textarea-item/';
+export { default as Toast } from './components/toast/';
+export { default as WhiteSpace } from './components/white-space/';
+export { default as WingBlank } from './components/wing-blank/';
+
+export { default as LocaleProvider } from './components/locale-provider/';
