+++ package.json
@@ -27,8 +27,8 @@
     "es"
   ],
   "license": "MIT",
-  "main": "lib/warn.js",
-  "module": "es/warn.js",
+  "main": "lib/index.js",
+  "module": "es/index.js",
   "dependencies": {
     "array-tree-filter": "~1.0.0",
     "babel-runtime": "6.x",
@@ -136,8 +136,8 @@
     "prepare": "antd-tools run guard",
     "prepublishOnly": "antd-tools run guard",
     "test": "npm run test:rn && npm run test:web",
-    "test:rn": "jest --config .jest.js",
-    "test:web": "jest --config .jest.web.js",
+    "test:rn": "jest --config .jest.native.js",
+    "test:web": "jest --config .jest.js",
     "test:all": "./scripts/test-all.sh",
     "authors": "git log --format='%aN <%aE>' | sort -u | grep -v 'users.noreply.github.com' > AUTHORS.txt",
     "dist": "BABEL_ENV=dist antd-tools run dist",
